requirements.txt
asgiref==3.6.0
Django==4.2
django-cuser==2017.3.16
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.18.9
django-tinymce==3.6.1
modeltranslation==0.25
Pillow==9.5.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3